活动管理和任务管理

+++ engine-dao/src/main/java/com/graduate/engine/model/Task.java
@@ -1,5 +1,6 @@
 package com.graduate.engine.model;
 
+import com.graduate.engine.model.viewobject.TreeData;
 import lombok.Data;
 
 @Data
@@ -22,7 +23,7 @@ public class Task {
 
     private Integer taskPriority;
 
-    private Integer taskStatusId;
+    private Long taskStatusId;
 
     private String taskMemo;
