+++ engine-dao/src/main/java/com/graduate/engine/model/TaskExec.java
@@ -12,7 +12,7 @@ public class TaskExec {
 
     private String taskExecName;
 
-    private Integer taskExecStatus;
+    private Long taskExecStatus;
 
     private String taskExecReport;
